requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
lightgbm>=3.3.0
opencv-python>=4.7.0
pytesseract>=0.3.10
flask>=2.2.0
streamlit>=1.20.0
langchain>=0.0.150
openai>=0.27.0
requests>=2.28.0
pytesseract>=0.3.10
google-generativeai>=0.7.0